Fix circular import issue
Moved ConstraintMap and ConstraintType in `src/jaxsim/api/kin_dyn_parameters.py`
