+++ src/jaxsim/api/kin_dyn_parameters.py
@@ -15,7 +15,6 @@ import jaxsim.typing as jtp
 from jaxsim.math import Inertia, JointModel, supported_joint_motion
 from jaxsim.math.adjoint import Adjoint
 from jaxsim.parsers.descriptions import JointDescription, JointType, ModelDescription
-from jaxsim.rbda.kinematic_constraints import ConstraintMap
 from jaxsim.utils import HashedNumpyArray, JaxsimDataclass
 
 
@@ -61,7 +60,7 @@ class KinDynParameters(JaxsimDataclass):
     hw_link_metadata: HwLinkMetadata | None = dataclasses.field(default=None)
 
     # Kinematic constraints
-    constraints: ConstraintMap
+    constraints: ConstraintMap | None = dataclasses.field(default=None)
 
     @property
     def motion_subspaces(self) -> jtp.Matrix:
@@ -1178,3 +1177,81 @@ class ScalingFactors(JaxsimDataclass):
 
     dims: jtp.Vector
     density: jtp.Float
+
+
+@dataclasses.dataclass(frozen=True)
+class ConstraintType:
+    """
+    Enumeration of all supported constraint types.
+    """
+
+    Weld: ClassVar[int] = 0
+    # TODO: handle Connect constraint
+    # Connect: ClassVar[int] = 1
+
+
+@jax_dataclasses.pytree_dataclass
+class ConstraintMap(JaxsimDataclass):
+    """
+    Class storing the kinematic constraints of a model.
+    """
+
+    frame_idxs_1: jtp.Int = dataclasses.field(
+        default_factory=lambda: jnp.array([], dtype=int)
+    )
+    frame_idxs_2: jtp.Int = dataclasses.field(
+        default_factory=lambda: jnp.array([], dtype=int)
+    )
+    constraint_types: jtp.Int = dataclasses.field(
+        default_factory=lambda: jnp.array([], dtype=int)
+    )
+    K_P: jtp.Float = dataclasses.field(
+        default_factory=lambda: jnp.array([], dtype=float)
+    )
+    K_D: jtp.Float = dataclasses.field(
+        default_factory=lambda: jnp.array([], dtype=float)
+    )
+
+    def add_constraint(
+        self,
+        frame_idx_1: int,
+        frame_idx_2: int,
+        constraint_type: int,
+        K_P: float | None = None,
+        K_D: float | None = None,
+    ) -> ConstraintMap:
+        """
+        Add a constraint to the constraint map.
+
+        Args:
+            frame_idx_1: The index of the first frame.
+            frame_idx_2: The index of the second frame.
+            constraint_type: The type of constraint.
+            K_P: The proportional gain for Baumgarte stabilization (default: 1000).
+            K_D: The derivative gain for Baumgarte stabilization (default: 2 * sqrt(K_P)).
+
+        Returns:
+            A new ConstraintMap instance with the added constraint.
+        """
+
+        # Set default values for Baumgarte coefficients if not provided
+        if K_P is None:
+            K_P = 1000
+        if K_D is None:
+            K_D = 2 * np.sqrt(K_P)
+
+        # Create new arrays with the input elements appended
+        new_frame_idxs_1 = jnp.append(self.frame_idxs_1, frame_idx_1)
+        new_frame_idxs2 = jnp.append(self.frame_idxs_2, frame_idx_2)
+        new_constraint_types = jnp.append(self.constraint_types, constraint_type)
+        new_K_P = jnp.append(self.K_P, K_P)
+        new_K_D = jnp.append(self.K_D, K_D)
+
+        # Return a new ConstraintMap object with updated attributes
+        return ConstraintMap(
+            frame_idxs_1=new_frame_idxs_1,
+            frame_idxs_2=new_frame_idxs2,
+            constraint_types=new_constraint_types,
+            K_P=new_K_P,
+            K_D=new_K_D,
+        )
